chore: move calculate search score to dapp explorer utils
